Added note on let's encrypt certs
Added a note explaining the location of let's encrypt certs with a link to the documentation. Let's Encrypt hasn't behaved as expected on RHEL/CentOS 7, in the sense that neither certbot or certbot-auto could be installed, either with yum, rpm or wget. The old letsencrypt script worked, though I'm not sure the certs were succesfully created.