Add: specified valid curves for ecdhe crypto
This was omitted before partly because the stronger curves weren't widely supported and partly because nginx only accepted a singel curve.

Currently, ssl_ecdh_curve acts exactly as ssl_ciphers: it offers the curves to the client out in the order they're specified, allowing only those on the list. I prefer Curve25519, with a fallback to NIST P-384, which is very widely supported. There are plenty of other suitable modern curves that could be added, but none that are widely supported enough to bother. Hopefully soon that will change.